Iteration 2 unit tests

See merge request comp3350s24/a01-01-flashcards!60

+++ app/src/main/java/comp3350/intellicards/Business/UserManager.java
@@ -32,7 +32,11 @@ public class UserManager {
         return null;
     }
 
-    public void deleteUser(String username) {
-        userPersistence.deleteUser(username);
+    public boolean deleteUser(String username) {
+        if (userPersistence.getUserByUsername(username) != null && !username.equals("guest")) {
+            userPersistence.deleteUser(username);
+            return true;
+        }
+        return false;
     }
 }
